refactor-update-logic
