+++ hazpy/flood/__init__.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+"""
+    hazpy
+    ~~~~~
+
+    FEMA developed module for analzying risk and loss from natural hazards.
+
+    :copyright: © 2019 by FEMA's Natural Hazards and Risk Assesment Program.
+    :license: cc, see LICENSE for more details.
+    :author: James Raines; james.rainesii@fema.dhs.gov
+"""
+
+__version__ = '0.0.1'
+__all__ = ['classes']
+
+from .classes import *
