+++ hazpy/flood/agriculture.py
@@ -1,0 +1,4 @@
+class Agriculture():
+    def __init__(self):
+        pass
+
