+++ hazpy/flood/direct_social_losses.py
@@ -1,0 +1,9 @@
+class DirectSocialLosses():
+    def __init__(self):
+        pass
+
+    def casualities(self):
+        pass
+
+    def shelter(self):
+        pass
