+++ hazpy/flood/indirect_economic_loss.py
@@ -1,0 +1,3 @@
+class IndirectEconomicLoss():
+    def __init__(self):
+        pass
