+++ hazpy/flood/modules/essential_facilities.py
@@ -1,0 +1,18 @@
+class EssentialFacilities():
+    def __init__(self):
+        pass
+
+    def medicalCare(self):
+        pass
+
+    def policeStations(self):
+        pass
+
+    def fireStations(self):
+        pass
+
+    def emergencyCenters(self):
+        pass
+
+    def schools(self):
+        pass
