+++ hazpy/flood/modules/general_building_stock.py
@@ -1,0 +1,18 @@
+class GeneralBuildingStock():
+    def __init__(self):
+        pass
+
+    def buildingAndContentDamage(self):
+        pass
+
+    def directEconomicLoss(self):
+        pass
+
+    def buildingDamageCount(self):
+        pass
+
+    def depreciatedBuildingAndContentLoss(self):
+        pass
+
+    
+
