+++ hazpy/flood/modules/user_defined_structures.py
@@ -1,0 +1,3 @@
+class UserDefinedStructures():
+    def __init__(self):
+        pass
