+++ hazpy/flood/modules/utility_systems.py
@@ -1,0 +1,54 @@
+# main class
+class UtilitySystems():
+    def __init__(self):
+        self.potableWater = PotableWater()
+        self.wasteWater = WasteWater()
+        self.oil = Oil()
+        self.naturalGas = NaturalGas()
+
+    def electricPower(self):
+        pass
+
+    def communication(self):
+        pass
+
+# helper classes
+class PotableWater():
+    def __init__(self):
+        pass
+
+    def facilities(self):
+        pass
+
+    def piplines(self):
+        pass
+
+class WasteWater():
+    def __init__(self):
+        pass
+    
+    def facilities(self):
+        pass
+
+    def piplines(self):
+        pass
+
+class Oil():
+    def __init__(self):
+        pass
+
+    def facilities(self):
+        pass
+
+    def piplines(self):
+        pass
+
+class NaturalGas():
+    def __init__(self):
+        pass
+
+    def facilities(self):
+        pass
+
+    def piplines(self):
+        pass
