+++ hazpy/flood/transportation_systems.py
@@ -1,0 +1,74 @@
+# main class
+class TransportationSystems():
+    def __init__(self):
+        self.highways = Highways()
+        self.railways = Railways()
+        self.lightrail = LightRail()
+        self.airport = AirportTransportation()
+
+    def busSystem(self):
+        pass
+
+    def portAndHarbor(self):
+        pass
+
+    def ferrySystem(self):
+        pass
+
+
+# helper classes
+class Highways():
+    def __init__(self):
+        pass
+
+    def roadways(self):
+        pass
+
+    def bridges(self):
+        pass
+
+    def tunnels(self):
+        pass
+
+class Railways():
+    def __init__(self):
+        pass
+
+    def tracksAndRoadbeds(self):
+        pass
+
+    def bridges(self):
+        pass
+
+    def tunnels(self):
+        pass
+
+    def systemFacilities(self):
+        pass
+
+class LightRail():
+    def __init__(self):
+        pass
+
+    def tracksAndRoadbeds(self):
+        pass
+
+    def bridges(self):
+        pass
+
+    def tunnels(self):
+        pass
+
+    def facilities(self):
+        pass
+
+class AirportTransportation():
+    def __init__(self):
+        pass
+
+    def facilities(self):
+        pass
+
+    def runways(self):
+        pass
+
