+++ hazpy/flood/vehicles.py
@@ -1,0 +1,3 @@
+class Vehicles():
+    def __init__(self):
+        pass
