+++ hazpy/flood/what_if.py
@@ -1,0 +1,21 @@
+class WhatIf():
+    def __init__(self):
+        pass
+
+    def includeFloodWarningParameters(self):
+        pass
+
+    def riverineLeveeAssessment(self):
+        pass
+
+    def riverineFlowRegulation(self):
+        pass
+
+    def riverineFloodVelocityAssessment(self):
+        pass
+
+    def coastalLongTermErosionAssessment(self):
+        pass
+
+    def coastalShoreProtectionAssessment(self):
+        pass
